+++ src/sections/contact/index.tsx
@@ -1,9 +1,9 @@
 import { SiMinutemailer, SiLinkedin, SiGithub, SiFacebook, SiInstagram } from "react-icons/si";
 
 import { handleCursorEnter, handleCursorLeave } from "@/utils/gsapUtils";
-import Section from "@/components/section/Section";
+import Section from "@/components/section";
 import { Button } from "@/components/ui/button";
-import SocialButton from "@/components/social-button/SocialButton";
+import SocialButton from "@/components/social-button";
 
 const socialLinks = [
   {
